+++ components/audio/include/ftl/audio/decoder.hpp
@@ -13,7 +13,7 @@ class Decoder {
 	Decoder() { };
 	virtual ~Decoder() { };
 
-	virtual bool decode(const ftl::codecs::Packet &pkt, std::vector<short> &out)=0;
+	virtual bool decode(const ftl::codecs::Packet &pkt, std::vector<float> &out)=0;
 
 	virtual bool accepts(const ftl::codecs::Packet &)=0;
 };
